Updated initConnection() and closeConnection()
These functions now utilize a "timeout" variable.  If you attempt to
connect to the DB and your connection fails, it will attempt to
reconnect.  The number of times the program attempts to reconnect is
determined by MAX_CONN_OPEN_ATTEMPTS.

If you attempt to close your DB conenction and it fails, it will attempt
to redisconnect.  The number of times the program attempts to disconnect
is determined by MAX_CONN_CLOSE_ATTEMPTS.

Both MAX_CONN_CLOSE_ATTEMPTS and MAX_CONN_OPEN_ATTEMPTS are constants
listed at the top of Tables.java